Utworzenie programu control_mod realizującego diagnostykę, wyłączanie czujników i zmianę okresu pomiarów
