+++ parameters.txt
@@ -1,3 +1,3 @@
-192.168.1.204 8080
-192.168.1.204 8081
-192.168.1.204 8082
+192.168.1.26 8080
+192.168.1.26 8081
+192.168.1.26 8082
