--- stop_sensors.h
@@ -1,8 +1,0 @@
-//stop_sensors.h
-//Autor: Magdalena Zych
-//Data: 11.05.2020
-
-#include <stdio.h>
-#include <string.h>
-#include "parameters.h"
-#include <arpa/inet.h> // inet_pton()
